delete kismina silmekten vazgec eklendi

+++ Otomasyon1.py
@@ -34,7 +34,7 @@ def list_student():
     os.system('cls')
     curser.execute("SELECT * FROM Students")
     students = curser.fetchall()
-    print("ID,Name,Surname,Age,Gender", sep="\t\t")
+    print("ID","Name","Surname","Age","Gender", sep="\t\t")
     print("------------------------------------------------------------------------------------------------------------------")
     for student in students:
         print(student[0],student[1],student[2],student[3],student[4], sep="\t\t")
@@ -90,33 +90,38 @@ def delete_student():
     os.system('cls')
     curser.execute("SELECT * FROM Students")
     students = curser.fetchall()
-    print("ID,Name,Surname,Age,Gender", sep="\t\t")
+    print("ID","Name","Surname","Age","Gender", sep="\t\t")
     print("------------------------------------------------------------------------------------------------------------------")
     for student in students:
         print(student[0],student[1],student[2],student[3],student[4], sep="\t\t")
     print("------------------------------------------------------------------------------------------------------------------")
-    valid_input = False
-    while not valid_input:
-        try: 
-            input_id = int(input("Enter id: "))
-            # Check if the ID exists in the database
-            if not id_exists_in_database(input_id):  # Replace with your database check logic
-                print("Invalid id. This ID does not exist. Please enter a valid ID.")
-            else:
-                valid_input = True
-                confirm = input("Öğrenciyi silmek istediğinize emin misiniz? (E/H): ")
-                if confirm.upper() == "E":
-                    curser.execute("DELETE FROM Students WHERE st_id = ?", (input_id,))
-                    conn.commit()
-                    print("Student deleted successfully.")
+    choice = (input("Press enter to continue or write '0' back to menu..."))
+    if choice == "0":
+        os.system('cls')
+        menu()
+    else:
+        valid_input = False
+        while not valid_input:
+            try: 
+                input_id = int(input("Enter id: "))
+                # Check if the ID exists in the database
+                if not id_exists_in_database(input_id):  # Replace with your database check logic
+                    print("Invalid id. This ID does not exist. Please enter a valid ID.")
                 else:
-                    print("Student not deleted.")
-        except ValueError:
-            print("Invalid id. Please enter a number.")
-
-    time.sleep(1)
-    os.system('cls')
-    menu()
+                    valid_input = True
+                    confirm = input("Öğrenciyi silmek istediğinize emin misiniz? (E/H): ")
+                    if confirm.upper() == "E":
+                        curser.execute("DELETE FROM Students WHERE st_id = ?", (input_id,))
+                        conn.commit()
+                        print("Student deleted successfully.")
+                        time.sleep(1)
+                        os.system('cls')
+                        menu()
+                    else:
+                        print("Student not deleted.")
+                        time.sleep(1)
+            except ValueError:
+                print("Invalid id. Please enter a number.")
 
 
 while option != 0:
